reorganize test package builder
